Get Network requests for vendors View is done
