+++ Clients/src/presentation/pages/Vendors/index.tsx
@@ -1,15 +1,20 @@
 import "./index.css";
 import { Button, Stack, Typography, useTheme } from "@mui/material";
-import TableWithPlaceholder from "../../components/Table/WithPlaceholder";
-import { useState } from "react";
+import TableWithPlaceholder from "../../components/Table/WithPlaceholder/index";
+import { useCallback, useContext, useEffect, useState } from "react";
 
 import AddNewVendor from "../../components/Modals/NewVendor";
 import singleTheme from "../../themes/v1SingleTheme";
+import { VerifyWiseContext } from "../../../application/contexts/VerifyWise.context";
+import { deleteEntityById, getAllEntities } from "../../../application/repository/entity.repository";
+import { logEngine } from "../../../application/tools/log.engine";
 
 const Vendors = () => {
   const theme = useTheme();
   const [isOpen, setIsOpen] = useState(false);
   const [value, setValue] = useState("1");
+  const [vendorChangeTrigger, setVendorChangeTrigger] = useState(0);
+  const { dashboardValues, setDashboardValues } = useContext(VerifyWiseContext);
 
   const openAddNewVendor = () => {
     setIsOpen(true);
@@ -18,6 +23,88 @@ const Vendors = () => {
   const handleChange = (_: React.SyntheticEvent, newValue: string) => {
     setValue(newValue);
   };
+
+  const fetchVendors = useCallback(async () => {
+    try {
+      const response = await getAllEntities({ routeUrl: "/vendors" });
+      console.log("response ===> ", response);
+      setDashboardValues((prevValues: any) => ({
+        ...prevValues,
+        vendors: response.data,
+      }));
+    } catch (error) {
+      console.error("Error fetching vendors:", error);
+    }
+  }, [setDashboardValues]);
+
+  useEffect(() => {
+    fetchVendors();
+  }, [fetchVendors, vendorChangeTrigger]);
+
+  const updateVendorChangeTrigger = () => {
+    setVendorChangeTrigger(prev => prev + 1);
+  };
+
+  // const handleEditVendor = async (vendorId: number, updatedVendorDetails: VendorDetails) => {
+  // try {
+  //   await editVendor(vendorId, updatedVendorDetails);
+  //   setVendorChangeTrigger(prev => prev + 1); // Increment to trigger re-fetch
+  // } catch (error) {
+  //   console.error("Error editing vendor:", error);
+  // }
+  // };
+
+  // const handleDeleteVendor = async (vendorId: number) => {
+  //   const user = {
+  //     id: "At delete vendor level", // Replace with actual user ID
+  //     email: "N/A", // Replace with actual user email
+  //     firstname: "N/A", // Replace with actual user first name
+  //     lastname: "N/A", // Replace with actual user last name
+  //   };
+
+  //   try {
+  //     const response = await deleteEntityById({
+  //       routeUrl: `/vendors/${vendorId}`,
+  //     });
+
+  //     if (response.status === 202) {
+  //       setDashboardValues((prevValues: any) => ({
+  //         ...prevValues,
+  //         vendors: prevValues.vendors.filter(
+  //           (vendor: any) => vendor.id !== vendorId
+  //         ),
+  //       }));
+  //       console.log("Vendor deleted successfully.");
+  //       logEngine({
+  //         type: "info",
+  //         message: "Vendor deleted successfully.",
+  //         user,
+  //       });
+  //       updateVendorChangeTrigger();
+  //     } else if (response.status === 404) {
+  //       console.error("Vendor not found.");
+  //       logEngine({
+  //         type: "error",
+  //         message: "Vendor not found.",
+  //         user,
+  //       });
+  //     } else {
+  //       console.error("Unexpected response. Please try again.");
+  //       logEngine({
+  //         type: "error",
+  //         message: "Unexpected response. Please try again.",
+  //         user,
+  //       });
+  //     }
+  //   } catch (error) {
+  //     console.error("Error deleting vendor:", error);
+  //     logEngine({
+  //       type: "error",
+  //       message: `An error occurred: ${error}`,
+  //       user,
+  //     });
+  //   }
+  // };
 
   return (
     <div className="vendors-page">
@@ -57,13 +144,19 @@ const Vendors = () => {
             Add new vendor
           </Button>
         </Stack>
-        <TableWithPlaceholder />
+        <TableWithPlaceholder
+          dashboardValues={dashboardValues}
+        // onVendorChange={updateVendorChangeTrigger}
+        // onDeleteVendor={handleDeleteVendor}
+        />
+
       </Stack>
       <AddNewVendor
         isOpen={isOpen}
         handleChange={handleChange}
         setIsOpen={() => setIsOpen(false)}
         value={value}
+      // onVendorChange={updateVendorChangeTrigger}
       />
     </div>
   );
